pyproject.toml
[tool.poetry]
name = "openapi2jsonschema"
version = "0.9.0"
description = "A utility to extract JSON Schema from a valid OpenAPI specification"
authors = ["Gareth Rushgrove <gareth@morethanseven.net>"]
license = "Apache-2.0"
readme = "README.md"
repository = "https://github.com/garethr/openapi2jsonschema"
keywords = ["openapi", "jsonschema"]

[tool.poetry.dependencies]
python = "^3.6"
jsonref = "^0.2.0"
pyyaml = "^5.1"
click = "^7.0"
colorama = "^0.4.1"

[tool.poetry.scripts]
openapi2jsonschema = "openapi2jsonschema.command:default"

[tool.poetry.dev-dependencies]
pytest = "^3.4"
pytest-black = "^0.3.2"
pytest-mypy = "^0.3.2"
pytest-cov = "^2.6"
pytest-datafiles = "^2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
